refactor: Share author-attribute derivation between the implicit line and entry paths
#1006 (issue #1003) landed `set_author_metadata` / `set_author_keys` in
`header.rs`, which derive the combined `authors` attribute and the `_N` / `_1`
companions from a resolved author list. That is exactly the logic this branch
had duplicated in `author_line.rs` to fix the implicit-author-line case.

Move `set_author_metadata` / `set_author_keys` into `document/author.rs` (next
to `Author`) as `pub(crate)` and call the shared helper from both
`AuthorLine::parse` (eagerly, so `{author}` / `{authors}` stay available to
later header lines) and `resolve_authors`. The bespoke `set_nth_attribute` and
`set_author_attributes` helpers are gone; there is now one derivation.

Also switch the affected `parser_test.rs` ports to a full `Parser::parse` so
they assert `authorcount` too, matching the Ruby originals, and correct the
now-stale "does not derive authorcount" notes (this crate does derive it; only
the author-less default of 0 is left unset).
