+++ parser/src/document/header.rs
@@ -4,7 +4,7 @@ use crate::{
     content::{Content, SubstitutionGroup, substitute_attributes_in_reftext},
     document::{
         Attribute, Author, AuthorLine, InterpretedValue, RefType, RevisionLine,
-        matches_author_pattern,
+        matches_author_pattern, set_author_metadata,
     },
     internal::{debug::DebugSliceReference, opaque_iter::opaque_slice_iter},
     span::MatchedItem,
@@ -1028,69 +1028,6 @@ fn split_author_entries(value: &str) -> Vec<&str> {
 
     entries.push(&value[start..]);
     entries
-}
-
-/// Populates the derived author document attributes from a resolved author
-/// list, mirroring Asciidoctor's `process_authors`.
-///
-/// The first author sets the unsuffixed keys (`author`, `firstname`,
-/// `authorinitials`, …); each subsequent author sets its `_N` companions. Once
-/// a second author appears, the first author is also mirrored onto its `_1`
-/// companions. The `authors` attribute is rewritten to the comma-joined list of
-/// resolved author names.
-///
-/// This intentionally does **not** honor `authorinitials_from_entry`: the
-/// derived `authorinitials` always overwrites an explicit `:authorinitials:`
-/// entry for the `authors` and `author_N` forms. Only a single `:author:` entry
-/// (handled inline in [`Header::parse`]) preserves an explicit override,
-/// exactly as Asciidoctor does – its `authorinitials` deletion guard lives only
-/// in the `author` branch of `process_authors`, not the `authors`/indexed
-/// branches.
-fn set_author_metadata(parser: &mut Parser, authors: &[Author]) {
-    for (idx, author) in authors.iter().enumerate() {
-        set_author_keys(parser, author, if idx == 0 { None } else { Some(idx + 1) });
-
-        // The `_1` companions are only assigned once a second author is seen.
-        if idx == 1
-            && let Some(first) = authors.first()
-        {
-            set_author_keys(parser, first, Some(1));
-        }
-    }
-
-    let joined = authors
-        .iter()
-        .map(Author::name)
-        .collect::<Vec<_>>()
-        .join(", ");
-
-    parser.set_attribute_by_value_from_header("authors", joined);
-}
-
-/// Sets the author attributes for a single author, either as the unsuffixed
-/// keys (`index` is `None`) or the `_N` companions (`index` is `Some(n)`).
-fn set_author_keys(parser: &mut Parser, author: &Author, index: Option<usize>) {
-    let key = |name: &str| match index {
-        None => name.to_string(),
-        Some(n) => format!("{name}_{n}"),
-    };
-
-    parser.set_attribute_by_value_from_header(key("author"), author.name());
-    parser.set_attribute_by_value_from_header(key("firstname"), author.firstname());
-
-    if let Some(middlename) = author.middlename() {
-        parser.set_attribute_by_value_from_header(key("middlename"), middlename);
-    }
-
-    if let Some(lastname) = author.lastname() {
-        parser.set_attribute_by_value_from_header(key("lastname"), lastname);
-    }
-
-    parser.set_attribute_by_value_from_header(key("authorinitials"), author.initials());
-
-    if let Some(email) = author.email() {
-        parser.set_attribute_by_value_from_header(key("email"), email);
-    }
 }
 
 fn apply_header_subs(source: &str, parser: &Parser) -> String {
