+++ parser/src/document/mod.rs
@@ -5,7 +5,7 @@ pub use attribute::{Attribute, InterpretedValue};
 
 mod author;
 pub use author::Author;
-pub(crate) use author::matches_author_pattern;
+pub(crate) use author::{matches_author_pattern, set_author_metadata};
 
 mod author_line;
 pub use author_line::{AuthorLine, Authors};
